python/requirements.txt
getpatter[local]
python-dotenv
httpx
